refactor: do not use `window.geolonia` as possible when parsing API key (rename)
