+++ src/lib/parse-atts.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import { keyring } from './parse-api-key';
+import { keyring } from './keyring';
 import { getLang } from './util';
 import type { EmbedAttributes } from '../embed';
